Moved FrictionPotential test
